fix: counter types to 256 bit

+++ src/libraries/ArrayMapping.sol
@@ -10,7 +10,7 @@ library ArrayMapping {
         returns (uint256[] memory)
     {
         uint256[] memory arrayOut = new uint256[](length);
-        for (uint8 i = 0; i < length; i++) {
+        for (uint256 i = 0; i < length; i++) {
             arrayOut[i] = _self[i];
         }
         return arrayOut;
@@ -20,7 +20,7 @@ library ArrayMapping {
      * @notice Set array values to mapping slots
      */
     function setValues(mapping(uint256 => uint256) storage _self, uint256[] calldata values) external {
-        for (uint8 i = 0; i < values.length; i++) {
+        for (uint256 i = 0; i < values.length; i++) {
             _self[i] = values[i];
         }
     }
